fix: show current image profile instead the previous one
