+++ src/pages/userProfile/styles.ts
@@ -4,7 +4,7 @@ const useStyles = makeStyles(() =>
   createStyles({
     root: {
       width: "100%",
-      height: "100%",
+      height: "80%",
       display: "flex",
       flexDirection: "column",
       alignItems: "center",
